perf c2c: Add dimensions for total LLC hit
Since Arm64 SPE trace data doesn't support HITM, we still want to
explore "perf c2c" tool to analyze cache false sharing.  If without HITM
tag, the tool cannot give out accurate result for cache false sharing,
a candidate solution is to sort the LLC hit and connect with the info of
multiple threads, e.g. if multiple threads hit the LLC on the same
cacheline for many times, this hints that it's likely to cause false
sharing issue.  Though this solution is not perfect due to lacking HITM
tag, it's pragmatic for detecting false sharing.

To support the sorting with LLC hit, this patch adds dimensions for
total LLC hit and the associated percentage calculation.

Signed-off-by: Leo Yan <leo.yan@linaro.org>