perf c2c: Add local variables for output metrics
For different display types, it might need to output different metrics
and sort on different metrics.

For this reason, this patch adds two local variables "output_str" and
"sort_str", one is used to point to output metrics and another points to
the sorting metrics; this is flexsible and the variables can be assigned
to different strings based on display type.

Signed-off-by: Leo Yan <leo.yan@linaro.org>