perf lock: Simplify adding thread stat
The tool uses RB tree to track the threads, it allocates structure
thread_stat if a new thread is coming, otherwise, finds the pointer for
the existed structure thread_stat by searching the RB tree.  Afterwards,
the structure thread_stat is used as a context to track all lock events
associated to a specific thread.

There have two functions thread_stat_findnew_first() and
thread_stat_findnew_after_first(), as the names indicate, the prior
one is used for the first thread, and the later one function is used for
later threads.  It's not necessary to use the two functions to maintain
the RB tree for thread stat, we can simply use one function for this:
for any threads (no matter if it's the first thread and a later thread),
search RB tree and directly use the existed thread stat structure,
otherwise allocates a new structure for the new coming thread.

This patch unifies functions as thread_stat_add() and is applied in the
whole file.

Signed-off-by: Leo Yan <leo.yan@linaro.org>