perf lock: Refactor code for adding lock stat
This patch refactors the code for adding the lock stat, it renames the
function lock_stat_findnew() to lock_stat_add(), so it's more clear to
say this function is used to add lock state node into hash table.

And in the function, it mixes the logic for memory allocation and
assignments, this patch uses the logic for firstly allocate memory and
then assign values to the data structure.

Signed-off-by: Leo Yan <leo.yan@linaro.org>