perf: lock: Use hlist for lock stat list
The tool needs to use a way for fast lockup for lock stats, but it uses
the general list for the hash table; this patch changes to use hlist
which is nature for managing hash table.

The comment clarifies the hash table is "based on kernel/lockdep.c", but
the hash table in the tool is no matter with any hash tables in lockdep;
so updates the comment to avoid confusion.

Signed-off-by: Leo Yan <leo.yan@linaro.org>