perf auxtrace: Introduce array to multiple PMU events
Currently, a single pointer is used to track a PMU event for AUX trace.
As a result, it is impossible to support multiple PMU events. On the
other hand, some Arm platforms classify CPUs into different Arm SPE PMU
events based on the CPU microarchitectures and topology. In this case,
only the AUX trace data from the first PMU event 'arm_pmu_0' can be
recorded.

This patch changes from the PMU pointer to a PMU pointer array in the
auxtrace_record structure, allowing it to save multiple PMU pointers.

Although it keeps to use the first PMU event, it is no functional
change. Based on it, later changes will support multiple PMU events for
AUX tracing.

Signed-off-by: Leo Yan <leo.yan@arm.com>