perf arm-spe: Support metadata version 2
This commit is to support metadata version 2 and at the meantime it is
backward compatible for version 1's format.

The metadata version 1 doesn't include the ARM_SPE_HEADER_VERSION field.
By checking the header size, it distinguishes the metadata is version 1
or version 2 (and any new versions if later will have). For version 2,
it reads out CPU number and retrieves the metadata info for every CPU.

Signed-off-by: Leo Yan <leo.yan@arm.com>