perf arm-spe: Use metadata to decide the data source feature
Use the info in the metadata to decide if the data source feature is
supported. The two factors are involving the decision, one is the CPU
MIDR which must be in the CPU list for the common data source encoding,
and another one is the Load Data Source (LDS) capacity.

For the metadata version 1, it doesn't include info for MIDR and LDS.
In this case, due to absent info for making decision, print out warning
to remind users to upgrade tool and returns false.

If the perf data contains the metadata version 2, only above two
conditions are met, it is safe to say the data source packet complies
the common encoding. Otherwise, it returns false.

Signed-off-by: Leo Yan <leo.yan@arm.com>