coresight: etm4x: Control the trace unit in CPU suspend
If a trace unit has been enabled, a CPU suspend operation misses to
disable the trace unit.  After the CPU resumes, it also does not follow
the general flow for re-enabling the trace unit.  As a result, this may
lead to a wrong state machine for the ETM.

This commit stops the trace unit during the CPU suspend flow and
re-enables the trace unit in the resume flow.

Fixes: f188b5e76aae ("coresight: etm4x: Save/restore state across CPU low power states")
Signed-off-by: Leo Yan <leo.yan@arm.com>