coresight: Set per CPU source pointer
Introduce coresight_set_percpu_source() for setting CPU source device. The
sources are maintained in a per CPU structure 'csdev_source'.

The coresight_register() function cannot be used for setting CPU source
device as it is absent info for CPU ID.  So this commit uses the ETM3 and
ETM4 drivers to set and clear CPU sources in probing and removal,
respectively.

Signed-off-by: Leo Yan <leo.yan@arm.com>