coresight: Manage activated paths during CPU hotplug
This commit handles activated paths during the CPU hotplug process.

When a CPU is hot-unplugged, and if an activated path is associated with
it, the CPU PM notifier disables the path.  This helps saving power,
especially when the CPU is expected to remain offline for long time.

Note, when disabling a path, then sink's disable() callback
automatically updates its buffer SYSFS mode.  Therefore, there is no
need to manually update the buffer.

During CPU hotplug-in, the previously active path is re-enabled
accordingly.

Signed-off-by: Leo Yan <leo.yan@arm.com>