coresight: Add PM callbacks in sink operation
Unlike a system level's sink, the per-CPU sink may lose power during CPU
idle states.  So far, this refers to the TRBE as a sink.  This commit
adds save and restore callbacks for the sink, and these callbacks are
via the PM notifier.

For not per-CPU sinks (e.g., ETR, ETB, etc), which are system level
components and should not be affected by CPU low power states.  The PM
notifier does not interact with them to avoid additional latency.

Signed-off-by: Leo Yan <leo.yan@arm.com>