coresight: Save activated path into source device
For a source device, save activated path into the coresight_device
structure. This will be used by later changes for controlling the path
during CPU idle.

Safe access to path pointers is guaranteed by the device mode - only
when the device is in an active mode. Therefore, the source device's
path pointer must be assigned before setting the device to an active
mode, and it must be cleared after setting the device to an inactive
mode.

Extract path enabling operations into a new _coresight_enable_path()
function, which will be used by CPU power management flow.

Signed-off-by: Leo Yan <leo.yan@arm.com>