coresight: Add per-CPU path pointer
Introduce per-CPU path pointers, save activated paths into the
structure. This will be used by later changes for controlling the path
during CPU idle.

The path pointer is assigned before setting the source device mode to
active, and it is cleared after the device is changed to an inactive
mode. So safe access to path pointers is guaranteed when the device is
in an active mode.

Refactor the path enabling logic into the _coresight_enable_path()
function. The existing coresight_enable_path() now calls this function
and updates the per-CPU pointers accordingly.

Signed-off-by: Leo Yan <leo.yan@arm.com>