coresight: etm4x: Remove redundant condition checks in save and restore
Since a dedicated callback is now used to determine whether context save
and restore operations are needed, performing the same check within the
save and restore callbacks is redundant.

The save and restore flows use two-level functions, where the first
level handles the condition check. With the removal of the redundant
checks, the two-level structure is no longer necessary. Polish the logic
into a single-level function.

Signed-off-by: Leo Yan <leo.yan@arm.com>