coresight: Manage activated paths during CPU hotplug
This commit handles activated paths during the CPU hotplug process.

When a CPU is hotplug off or hotplug in, and if an activated path is
associated with it, the CPU PM notifier disables and enables the path,
including the sink and helpers.

When disabling a path, the sink's disable() callback updates its buffer
in SysFS mode. No need to manually update the buffer in this case.

Signed-off-by: Leo Yan <leo.yan@arm.com>