coresight: cti: Introduce a new mode CS_MODE_DEBUG
Introduce a new CS_MODE_DEBUG mode to indicate that a device is being
used for debugging purposes, e.g, if a system ignores CoreSight trace
functionality and enables CTI for debugging.

The CTI driver retains the same behavior for the CS_MODE_DEBUG mode. For
the SysFS and Perf modes, this change is a preparation for future
updates where the core layer will control the path (including helpers
like CTI devices) in CPU power management. The CTI driver keeps to
update the hw_powered flag to correctly reflect the power state.

Update ETM4x driver to mute compiler warnings for the newly added mode.

Signed-off-by: Leo Yan <leo.yan@arm.com>