coresight: Set per-CPU source pointer
Add coresight_{set|clear|get}_percpu_source() helpers to access the
per-CPU source pointer.

Device registration and unregistration are first consumers to call the
set and clear helpers.

For the set or clear helper, if is called from a CPU hotplug notifier
(see etm4_probe_cpu()), it sets the pointer directly since it is already
running on the local CPU.  Otherwise, to avoid races with CPU PM
notifiers, it acquires the CPU lock and issues an SMP call to update
the per-CPU pointer on the target CPU.  If that fails, it sets the
pointer as a last resort.

The helper coresight_get_percpu_source() is added for future changes
to retrieve the pointer.

Tested-by: James Clark <james.clark@linaro.org>
Reviewed-by: Yeoreum Yun <yeoreum.yun@arm.com>
Reviewed-by: James Clark <james.clark@linaro.org>
Signed-off-by: Leo Yan <leo.yan@arm.com>