coresight: trbe: Remove misleading comment
The comment describes two scenarios where the calculated limit is less
than the head.

However, since the wakeup address is always higher than the head, a
round-up wakeup must also be higher than the head. Therefore, it is not
possible for the limit to be capped to a value less than the head due to
a wakeup.

As a result, the two scenarios effectively describe the same condition.
Although they differ in whether a wakeup is involved, the wakeup has no
impact on the final outcome. Thus, the second scenario is redundant.

Remove the comment to avoid confusion.

Signed-off-by: Leo Yan <leo.yan@arm.com>