coresight: trbe: Support trigger count
Currently, if the watermark is before the buffer's wrap point, the limit
is capped at the watermark:

      head            watermark tail
 +----+---------------+---------+-------+
 |****|               |         |*******|
 +----+---------------+---------+-------+
 base                 limit             base + nr_pages

 *** : Filled trace data

The buffer operates in fill mode, where tracing stops when reaching the
end of the buffer and a maintenance interrupt is raised. However, due to
IRQ latency, trace data may be lost during the window where tracing is
halted but the program continues to run.

To mitigate the issue, this commit uses the trigger count to replace the
limit for watermark. It sets trigger count as a delta between the head
and the watermark:

      head            watermark tail
 +----+---------------+---------+-------+
 |****|               |         |*******|
 +----+---------------+---------+-------+
 base `---- count ----'         limit   base + nr_pages

 *** : Filled trace data

Once the count is decremented to 0, an interrupt is asserted, but the
hardware continues collecting trace data. This effectively prevents
trace loss.

Fill mode is still used, it stops the hardware to prevent overwriting
in case the software cannot read the data in time.

Signed-off-by: Leo Yan <leo.yan@arm.com>